requirements.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.2
Pygments==2.4.2
pylint==2.4.3
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
wrapt==1.11.2
